Service - Move files for namespace consistency
