+++ TF_Service_dotNet/TouchFree/Connections/IClientConnectionManager.cs
@@ -1,5 +1,5 @@
 ﻿using System.Net.WebSockets;
-using Ultraleap.TouchFree.Library.Connection;
+using Ultraleap.TouchFree.Library.Connections;
 
 namespace Ultraleap.TouchFree.Library
 {
