+++ TF_Service_dotNet/TouchFree/Connections/MessageQueues/ConfigurationChangeQueueHandler.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json;
 using Ultraleap.TouchFree.Library.Configuration;
 
-namespace Ultraleap.TouchFree.Library.Connection
+namespace Ultraleap.TouchFree.Library.Connections.MessageQueues
 {
     public class ConfigurationChangeQueueHandler : BaseConfigurationChangeQueueHandler
     {
