+++ TF_Service_dotNet/TouchFree/Connections/MessageQueues/HandDataStreamStateQueueHandler.cs
@@ -1,8 +1,7 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using Ultraleap.TouchFree.Library.Connections;
 
-namespace Ultraleap.TouchFree.Library.Connection
+namespace Ultraleap.TouchFree.Library.Connections.MessageQueues
 {
     public class HandDataStreamStateQueueHandler : MessageQueueHandler
     {
