+++ TF_Service_dotNet/TouchFree/Connections/MessageQueues/IMessageQueueHandler.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Concurrent;
 
-namespace Ultraleap.TouchFree.Library.Connection
+namespace Ultraleap.TouchFree.Library.Connections.MessageQueues
 {
     public interface IMessageQueueHandler
     {
