+++ TF_Service_dotNet/TouchFree/Connections/MessageQueues/MessageQueueHandler.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json.Linq;
 using System.Collections.Concurrent;
 
-namespace Ultraleap.TouchFree.Library.Connection
+namespace Ultraleap.TouchFree.Library.Connections.MessageQueues
 {
     public abstract class MessageQueueHandler : IMessageQueueHandler
     {
