+++ TF_Service_dotNet/TouchFree_Service/Connection/ClientConnection.cs
@@ -9,7 +9,8 @@ using System.Text.RegularExpressions;
 using System.Threading;
 using Ultraleap.TouchFree.Library;
 using Ultraleap.TouchFree.Library.Configuration;
-using Ultraleap.TouchFree.Library.Connection;
+using Ultraleap.TouchFree.Library.Connections;
+using Ultraleap.TouchFree.Library.Connections.MessageQueues;
 using Ultraleap.TouchFree.Service.ConnectionTypes;
 
 namespace Ultraleap.TouchFree.Service.Connection
