+++ TF_Service_dotNet/TouchFree_Service/Connection/Middlewares/TouchFreeRouter.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Ultraleap.TouchFree.Library;
 using Ultraleap.TouchFree.Library.Configuration;
-using Ultraleap.TouchFree.Library.Connection;
+using Ultraleap.TouchFree.Library.Connections.MessageQueues;
 
 namespace Ultraleap.TouchFree.Service.Connection
 {
