+++ TF_Service_dotNet/TouchFree_Service/Connection/Middlewares/TouchFreeRouterExtensions.cs
@@ -6,9 +6,9 @@ using Ultraleap.TouchFree.Library;
 using Ultraleap.TouchFree.Library.Connections;
 using Ultraleap.TouchFree.Library.Configuration;
 using Ultraleap.TouchFree.Library.Configuration.QuickSetup;
-using Ultraleap.TouchFree.Library.Connection;
 using Ultraleap.TouchFree.Library.Interactions;
 using Ultraleap.TouchFree.Library.Interactions.PositionTrackers;
+using Ultraleap.TouchFree.Library.Connections.MessageQueues;
 
 namespace Ultraleap.TouchFree.Service.Connection
 {
